added react router, cleaned up App.tsx, added a page template to follow for all pages
