+++ client/package.json
@@ -25,6 +25,7 @@
     "react-dom": "^18.2.0",
     "react-material-ui-carousel": "^3.4.2",
     "react-redux": "^8.0.5",
+    "react-router-dom": "^6.11.0",
     "react-scripts": "5.0.1",
     "recharts": "^2.1.14",
     "sass": "^1.55.0",
